refactor: RacingIO 폴더경로 수정 domain -> view
